Moved ks statistic into univariate
