+++ copulas/univariate/__init__.py
@@ -1,4 +1,5 @@
-from copulas.metrics import ks_statistic
+import numpy as np
+
 from copulas.univariate.base import Univariate
 from copulas.univariate.beta import BetaUnivariate
 from copulas.univariate.gaussian import GaussianUnivariate
@@ -14,6 +15,14 @@ __all__ = (
     'TruncatedGaussian',
     'Univariate',
 )
+
+
+def ks_statistic(copula, X):
+    copula.fit(X)
+    estimated_cdf = np.sort(copula.cumulative_distribution(X))
+    emperical_cdf = np.linspace(0.0, 1.0, num=len(X))
+    statistic = max(np.abs(estimated_cdf - emperical_cdf))
+    return statistic
 
 
 def select_univariate(X):
