+++ tests/copulas/univariate/test_select.py
@@ -3,12 +3,11 @@ from unittest import TestCase
 import numpy as np
 from scipy.stats import truncnorm
 
-from copulas.metrics import ks_statistic
 from copulas.univariate import (
-    GaussianKDE, GaussianUnivariate, TruncatedGaussian, select_univariate)
+    GaussianKDE, GaussianUnivariate, TruncatedGaussian, ks_statistic, select_univariate)
 
 
-class TestKSStatistic(TestCase):
+class TestSelectUnivariate(TestCase):
 
     def setUp(self):
         size = 1000
